Uploading updated User class
This commit uploads the new User class that implements our new class structure

+++ backend/User/UserAccessor.cs
@@ -1,32 +1,42 @@
-public class UserAccessor : IUserAccessor
+namespace User
 {
-    private IUserRepository _userRepository;
-
-    public UserAccessor(IUserRepository userRepository)
+    public class UserAccessor : IUserAccessor
     {
-        _userRepository = userRepository;
-    }
+        private readonly IUserRepository _userRepository;
 
-    public User GetUserById(int userId)
-    {
-        return _userRepository.GetUserById(userId);
-    }
+        public UserAccessor(IUserRepository userRepository)
+        {
+            _userRepository = userRepository;
+        }
 
-    public void CreateUser(User user)
-    {
-        _userRepository.CreateUser(user);
-        return;
-    }
+        public User GetUserById(int userId)
+        {
+            return _userRepository.GetUserById(userId);
+        }
 
-    public void UpdateUser(int id, User user)
-    {
-        _userRepository.UpdateUser(id, user);
-        return;
-    }
+        public bool CreateUser(User user)
+        {
+            return _userRepository.CreateUser(user);
+        }
 
-    public void DeleteUser(int id)
-    {
-        _userRepository.DeleteUser(id);
-        return;
+        public bool UpdateUser(int id, User user)
+        {
+            var tempuser = _userRepository.GetUserById(id);
+            if (tempuser == null)
+            {
+                throw new ArgumentException($"No user found with ID {id}");
+            }
+            return _userRepository.UpdateUser(id, user); ;
+        }
+
+        public bool DeleteUser(int id)
+        {
+            var tempuser = _userRepository.GetUserById(id);
+            if (tempuser == null)
+            {
+                throw new ArgumentException($"No user found with ID {id}");
+            }
+            return _userRepository.DeleteUser(id); ;
+        }
     }
 }
